requirements.txt
streamlit
numpy
Pillow
scikit-image
torch==2.7.0
torchvision==0.16.0
opencv-python-headless==4.10.0.84
